Fixes #38: Update to support wct 6.4.0+ & polyserve
When `wct` moved from `serve-waterfall` to `polyserve`, it broke
this plugin's usage of the `prepare:webserver` hook.  With newer
versions of `wct`, the app used by `wct` to serve component files
has a wildcard request handler registered that prevents calling
the instrumentation middleware.

Version 6.4.0 of `wct` adds a new `define:webserver` hook so that
plugins can inject middleware to be executed before the `polyserve`
wildcard handler.

This replaces the `prepare:webserver` handler with one for the new
`define:webserver` event, allowing this plugin to instrument code
on the fly again.

Also fixes #41